feat: 🎸 tailwind表示されないバクの更新
