+++ frontend/app/types/ReactFetch.d.ts
@@ -1,0 +1,1 @@
+declare module 'react-fetch';
