+++ frontend/components/BranchDiagram.tsx
