+++ frontend/next.config.js
@@ -1,9 +1,4 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {}
 
-module.exports = nextConfig;
-module.exports = {
-  images: {
-    domains: ["camo.qiitausercontent.com"],
-  },
-};
+module.exports = nextConfig
