--- frontend/types/ReactFetch.d.ts
@@ -1,1 +1,0 @@
-declare module 'react-fetch';
